refactor disciples functions
